Add abstract base class for single writer trade loggers
This contains the common logic shared by both the CSV and the new SQLITE
trade loggers now.

Add general handling of asynchronous setups: While the setup is still in
progress, we buffer any incoming trade records and log them once the setup
completes. If the setup failed, we disable the trade logger.
During shutdown/flush, we await the setup task before we save the pending
trades.

These changes improve the sqlite trade logger:
* Buffer trades while the setup is in progress.
* Only use a single concurrent writer to write to the sqlite database
  (sqlite only supports a single writer).
* Buffer incoming trades over a short time period to reduce IO overhead.
* Avoid using BukkitScheduler.RunTaskAsynchronously directly: There are
  certain edge cases (plugin shutdowns) that one needs to be aware of (see
  SchedulerUtils).
* Retry failed save attempts.
  Ignore (but still log) SQLExceptions caught during connection closing:
  Since the previous inserts completed successfully, we assume that the
  trades were successfully saved (no need to retry the saving of those).
