+++ modules/main/src/main/java/com/nisovin/shopkeepers/tradelog/base/package-info.java
@@ -1,0 +1,2 @@
+@org.eclipse.jdt.annotation.NonNullByDefault
+package com.nisovin.shopkeepers.tradelog.base;
